csv_uploader_backend/requirements.txt
django==5.0.4
django-cors-headers==4.3.0
djangorestframework==0.1.0
gunicorn==21.2.0